refactor: #82 코드리뷰 적용
